+++ zcash_client_sqlite/src/lib.rs
@@ -458,7 +458,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
         protocol: ShieldedProtocol,
         index: u32,
         target_height: TargetHeight,
-        _include_locked: bool,
+        include_locked: bool,
     ) -> Result<Option<ReceivedNote<Self::NoteRef, Note>>, Self::Error> {
         match protocol {
             ShieldedProtocol::Sapling => wallet::sapling::get_spendable_sapling_note(
@@ -467,6 +467,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                 txid,
                 index,
                 target_height,
+                include_locked,
             )
             .map(|opt| opt.map(|n| n.map_note(Note::Sapling))),
             ShieldedProtocol::Orchard => {
@@ -477,6 +478,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                     txid,
                     index,
                     target_height,
+                    include_locked,
                 )
                 .map(|opt| opt.map(|n| n.map_note(Note::Orchard)));
 
@@ -494,7 +496,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
         target_height: TargetHeight,
         confirmations_policy: ConfirmationsPolicy,
         exclude: &[Self::NoteRef],
-        _include_locked: bool,
+        include_locked: bool,
     ) -> Result<ReceivedNotes<Self::NoteRef>, Self::Error> {
         Ok(ReceivedNotes::new(
             if sources.contains(&ShieldedProtocol::Sapling) {
@@ -506,6 +508,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                     target_height,
                     confirmations_policy,
                     exclude,
+                    include_locked,
                 )?
             } else {
                 vec![]
@@ -520,6 +523,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                     target_height,
                     confirmations_policy,
                     exclude,
+                    include_locked,
                 )?
             } else {
                 vec![]
@@ -533,7 +537,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
         sources: &[ShieldedProtocol],
         target_height: TargetHeight,
         exclude: &[Self::NoteRef],
-        _include_locked: bool,
+        include_locked: bool,
     ) -> Result<ReceivedNotes<Self::NoteRef>, Self::Error> {
         Ok(ReceivedNotes::new(
             if sources.contains(&ShieldedProtocol::Sapling) {
@@ -547,6 +551,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                     ShieldedProtocol::Sapling,
                     wallet::sapling::to_received_note,
                     wallet::common::NoteRequest::Unspent,
+                    include_locked,
                 )?
             } else {
                 vec![]
@@ -563,6 +568,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
                     ShieldedProtocol::Orchard,
                     wallet::orchard::to_received_note,
                     wallet::common::NoteRequest::Unspent,
+                    include_locked,
                 )?
             } else {
                 vec![]
@@ -590,7 +596,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
         address: &TransparentAddress,
         target_height: TargetHeight,
         confirmations_policy: ConfirmationsPolicy,
-        _include_locked: bool,
+        include_locked: bool,
     ) -> Result<Vec<WalletUtxo>, Self::Error> {
         wallet::transparent::get_spendable_transparent_outputs(
             self.conn.borrow(),
@@ -598,6 +604,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
             address,
             target_height,
             confirmations_policy,
+            include_locked,
         )
     }
 
@@ -608,7 +615,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
         selector: &NoteFilter,
         target_height: TargetHeight,
         exclude: &[Self::NoteRef],
-        _include_locked: bool,
+        include_locked: bool,
     ) -> Result<AccountMeta, Self::Error> {
         let sapling_pool_meta = unspent_notes_meta(
             self.conn.borrow(),
@@ -617,6 +624,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
             account_id,
             selector,
             exclude,
+            include_locked,
         )?;
 
         #[cfg(feature = "orchard")]
@@ -627,6 +635,7 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters, CL, R> InputSour
             account_id,
             selector,
             exclude,
+            include_locked,
         )?;
         #[cfg(not(feature = "orchard"))]
         let orchard_pool_meta = None;